refactor markdown parsing into its own class
