+++ genai-engine/src/utils/utils.py
@@ -3,13 +3,11 @@ import functools
 import logging
 import os
 import re
-import string
 import traceback
 import urllib
 from concurrent.futures import ThreadPoolExecutor
 from typing import Callable, List, Union
 
-import commonmark
 import utils.constants as constants
 from dotenv import load_dotenv
 from fastapi import HTTPException, Query
@@ -20,7 +18,6 @@ from opentelemetry.sdk.trace import Tracer
 from schemas.common_schemas import LLMTokenConsumption, PaginationParameters
 from schemas.enums import PaginationSortMethod
 from sqlalchemy.orm import Session
-from utils.abbreviations import ABBREVIATIONS
 
 _root_dir = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
 _genai_engine_version = None
@@ -224,105 +221,6 @@ async def common_pagination_parameters(
     if page_size > constants.MAX_PAGE_SIZE or page_size <= 0:
         raise HTTPException(status_code=400, detail=constants.ERROR_PAGE_SIZE_TOO_LARGE)
     return PaginationParameters(sort=sort, page_size=page_size, page=page)
-
-
-def deduplicate(seq: list[str]) -> list[str]:
-    """
-    Source: https://stackoverflow.com/a/480227/1493011
-    """
-
-    seen = set()
-    return [x for x in seq if not (x in seen or seen.add(x))]
-
-
-def strip_markdown(text: str) -> str:
-    """
-    Strip Markdown from a LLM Response
-    """
-    parser = commonmark.Parser()
-
-    def ast2text(astNode):
-        """
-        Returns the text from markdown, stripped of the markdown syntax itself
-        """
-        walker = astNode.walker()
-        acc = ""
-        iterator = iter(walker)
-        list_level = 0
-        for current, entering in iterator:
-            if current.literal and not (
-                current.parent
-                and current.parent.t == "link"
-                and current.parent.destination == current.literal
-            ):
-                acc += current.literal
-            if current.t == "linebreak":
-                acc += "\n"
-            elif current.t == "softbreak":
-                acc += " "
-            elif current.t == "list" and entering:
-                if list_level > 0:
-                    # Already in a list
-                    acc = acc.strip() + " "
-                    # Sub the last new line, the rest of the item is supposed to be on the same line
-                list_level += 1
-            elif current.t == "list" and not entering:
-                list_level -= 1
-                if list_level <= 1:
-                    acc = acc.strip()
-                    acc += "\n"
-            elif current.t == "paragraph" and not entering:
-                if list_level > 1:
-                    if acc[-1] in string.punctuation:
-                        acc = acc[:-1]  # Strip punctuation for list items
-                    acc += " "  # Don't add new line until exiting nested alist
-                else:
-                    acc = acc.strip()
-                    acc += "\n"
-            elif current.t == "heading" and not entering:
-                acc += " "
-            elif current.t in ("link", "image") and entering is False:
-                acc += f" {current.destination}"
-                if current.title:
-                    acc += f" - {current.title}"
-        return acc.strip()
-
-    try:
-        ast = parser.parse(text)
-        parsed = ast2text(ast)
-    except Exception as e:
-        parsed = text
-        logger.warning(f"Failed to parse text with exception {e}")
-
-    return parsed
-
-
-def custom_text_parser(text, delims="all") -> list[str]:
-    """
-    Returns a list of texts that should contain sentences & list items from an LLM response
-    """
-    text = strip_markdown(text)
-    abbreviation_pattern = r"([A-Za-z]\.)([A-Za-z]\.)+"
-    all_abbreviations = re.finditer(abbreviation_pattern, text)
-
-    # Iterate through all found emails and replace .com
-    for abbrev in all_abbreviations:
-        found = abbrev.group(0)
-        text = text.replace(found, found.replace(".", ""))
-
-    for s in ABBREVIATIONS:
-        text = text.replace(s, s.replace(".", ""))
-
-    lines = text.strip().split("\n")
-    texts = []
-    for line in lines:
-        line = line.strip()
-        if list_indicator_regex.match(line.strip()):
-            texts.append(line.strip())
-        else:
-            texts.extend(sentence_tokenizer.tokenize(line))
-    return deduplicate(texts)
-
 
 def pad_text(
     text: Union[str, List[str]],
